+++ tests/test_general.py
@@ -9,7 +9,7 @@ import pytest
 
 
 def test_version():
-    assert "1.5.0" == deepl.__version__
+    assert "1.7.0" == deepl.__version__
 
 
 @pytest.mark.parametrize(
@@ -23,8 +23,9 @@ def test_example_translation(lang, translator):
     The texts are translations of "proton beam"."""
 
     input_text = example_text[lang]
+    source_lang = deepl.Language.remove_regional_variant(lang)
     result_text = translator.translate_text(
-        input_text, target_lang="EN-US"
+        input_text, source_lang=source_lang, target_lang="EN-US"
     ).text.lower()
     assert "proton" in result_text
 
@@ -145,6 +146,11 @@ def test_usage_overrun(translator_with_random_auth_key, server, tmpdir):
     )
 
     usage = translator.get_usage()
+    assert usage.any_limit_reached
+    assert usage.document.limit_reached
+    assert usage.character.limit_reached
+    assert not usage.team_document.limit_reached
+    # Test deprecated properties as well
     assert usage.any_limit_exceeded
     assert usage.document.limit_exceeded
     assert usage.character.limit_exceeded
@@ -170,7 +176,7 @@ def test_usage_team_document_limit(
 
     translator = translator_with_random_auth_key
     usage = translator.get_usage()
-    assert not usage.any_limit_exceeded
+    assert not usage.any_limit_reached
     assert "Characters" not in str(usage)
     assert "Documents" not in str(usage)
     assert "Team documents: 0 of 1" in str(usage)
@@ -184,7 +190,7 @@ def test_usage_team_document_limit(
     )
 
     usage = translator.get_usage()
-    assert usage.any_limit_exceeded
-    assert not usage.document.limit_exceeded
-    assert not usage.character.limit_exceeded
-    assert usage.team_document.limit_exceeded
+    assert usage.any_limit_reached
+    assert not usage.document.limit_reached
+    assert not usage.character.limit_reached
+    assert usage.team_document.limit_reached